# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aharraa API",
    "version": "0.1.0",
    "description": "Auto-generated OpenAPI spec for Aharraa server"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user using Supabase email/password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signup successful (may require email confirmation)",
            "headers": {
              "X-Access-Token": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase access token"
              },
              "X-User-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase user ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "summary": "Sign in with email/password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signin successful; returns Supabase session info",
            "headers": {
              "X-Access-Token": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase access token"
              },
              "X-User-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase user ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (validation error or invalid credentials)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/oauth/google": {
      "get": {
        "summary": "Get a Supabase OAuth redirect URL for Google",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "redirect",
            "schema": {
              "type": "string"
            },
            "description": "Optional redirect URL after OAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a URL to redirect the user to start Google OAuth"
          }
        }
      }
    },
    "/auth/callback": {
      "get": {
        "summary": "OAuth callback endpoint for Supabase redirects",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "access_token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Access token returned by Supabase after OAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth callback successful",
            "headers": {
              "X-Access-Token": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase access token"
              },
              "X-User-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Supabase user ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Clear server session (logout)",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Session cleared"
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "summary": "Verify a Supabase access token and return user info",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "access_token"
                ],
                "properties": {
                  "access_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token valid and user info returned"
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Create a new order in the system",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "items",
                  "shippingAddress",
                  "billingAddress",
                  "paymentMethod",
                  "totalAmount",
                  "currency"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "The ID of the user placing the order"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "productId",
                        "quantity",
                        "price"
                      ],
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "The ID of the product (meal)"
                        },
                        "quantity": {
                          "type": "number",
                          "description": "The quantity of this product"
                        },
                        "price": {
                          "type": "number",
                          "description": "The price of this specific item (quantity * meal price * plan duration)"
                        }
                      }
                    }
                  },
                  "shippingAddress": {
                    "type": "object",
                    "required": [
                      "street",
                      "city",
                      "zip"
                    ],
                    "properties": {
                      "street": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "zip": {
                        "type": "string"
                      }
                    }
                  },
                  "billingAddress": {
                    "type": "object",
                    "required": [
                      "street",
                      "city",
                      "zip"
                    ],
                    "properties": {
                      "street": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "zip": {
                        "type": "string"
                      }
                    }
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "e.g., \"COD\" (Cash On Delivery), \"credit_card\", \"paypal\""
                  },
                  "totalAmount": {
                    "type": "number",
                    "format": "float",
                    "description": "The total amount of the order"
                  },
                  "currency": {
                    "type": "string",
                    "description": "The currency of the total amount, e.g., \"INR\""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The unique ID of the newly created order"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "productId": {
                            "type": "string"
                          },
                          "quantity": {
                            "type": "number"
                          },
                          "price": {
                            "type": "number"
                          },
                          "mealName": {
                            "type": "string"
                          },
                          "planName": {
                            "type": "string"
                          },
                          "vendorName": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "shippingAddress": {
                      "type": "object",
                      "properties": {
                        "street": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string"
                        },
                        "zip": {
                          "type": "string"
                        }
                      }
                    },
                    "billingAddress": {
                      "type": "object",
                      "properties": {
                        "street": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string"
                        },
                        "zip": {
                          "type": "string"
                        }
                      }
                    },
                    "paymentMethod": {
                      "type": "string"
                    },
                    "totalAmount": {
                      "type": "number",
                      "format": "float"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "orderDate": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO date string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "confirmed",
                        "delivered",
                        "cancelled"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Orders",
      "description": "Operations related to orders"
    }
  ]
}